+++ js/app/dashboard/redux/__tests__/taskEntityReducers.test.js
@@ -1,4 +1,4 @@
-import { default as taskEntityReducers } from '../redux/taskEntityReducers'
+import { default as taskEntityReducers } from '../taskEntityReducers'
 
 describe('taskEntityReducers', () => {
